KURWA

+++ julek.py
@@ -119,9 +119,9 @@ def europe(SIR0 = None):
     num = dict(zip(Labels, range(len(Labels))))
     A = pd.read_csv("backend/thematrix.csv" , header = None).values/(1000)
 
-    df = pd.read_csv("backend/SIR0.csv").astype(float)
+    df = pd.read_csv("backend/SIR0.csv")
     df = df.loc[df["Month"] == 3].loc[df["Day"] == 12]
-    SIR0 = np.array([N]+[[0]*len(N)]*2)
+    SIR0 = np.array([N]+[[0]*len(N)]*2).astype(float)
     for i in Labels:
         SIR0[1,num[i]] = df.loc[df["Label"] == i,"Cases"].iloc()[0]/1000
         SIR0[0,num[i]] -= SIR0[1,num[i]]
